+++ src/utils/index.ts
@@ -3,8 +3,8 @@
  * 统一导出所有工具函数，便于模块化导入
  */
 
-export * from './sqlBeautifier'
-export * from './operationLogger'
-export * from './sqlStats'
-export * from './fileUploader'
-export * from './fieldMapping'
+export * from './sql/sqlBeautifier'
+export * from './log/operationLogger'
+export * from './sql/sqlStats'
+export * from './file/fileUploader'
+export * from './field/fieldMapping'